Description
To remove the noise from the noise signal, it is significant to analyze the noise signal in time domain and frequency domain. While time-domain analyze a signal changes over time, and frequency domain presents the distribution of signal’s energy. Also the majority of frequency component can be obtained. In order to do this, Fast Fourier transform (FFT) is the quickest way to transform from time domain to frequency domain.